consolidating files to reduce project sprawl
