+++ cmd/referee-lights/main.go
@@ -3,19 +3,296 @@
 package main
 
 import (
-	"github.com/aws/aws-xray-sdk-go/xray"
-	"github.com/joho/godotenv"
-	"go-ref-lights/internal/app"
+	"fmt"
+	"html/template"
+	"io"
 	"log"
 	"net/http"
 	"os"
+	"path/filepath"
+	"runtime"
+	"strings"
+	"sync"
 	"time"
 
+	"github.com/aws/aws-xray-sdk-go/xray"
+	"github.com/gin-contrib/sessions"
+	"github.com/gin-contrib/sessions/cookie"
+	"github.com/gin-gonic/gin"
+	"github.com/joho/godotenv"
+
 	"go-ref-lights/controllers"
-	"go-ref-lights/heartbeat"
 	"go-ref-lights/logger"
+	"go-ref-lights/middleware"
+	"go-ref-lights/services"
 	"go-ref-lights/websocket"
 )
+
+// SetupRouter creates and configures a Gin router.
+func SetupRouter(env string) *gin.Engine {
+	// configure Gin mode
+	if env == "production" {
+		gin.SetMode(gin.ReleaseMode)
+	} else {
+		gin.SetMode(gin.TestMode)
+	}
+
+	router := gin.Default()
+
+	// serve static files
+	router.Static("/static", "./static")
+	router.StaticFile("/favicon.ico", "./static/images/favicon.ico")
+
+	// reduce logs in non-production
+	if env != "production" {
+		gin.DefaultWriter = io.Discard
+		gin.DefaultErrorWriter = io.Discard
+		logger.Debug.Println("[SetupRouter] Gin logs have been discarded for non-production mode.")
+	}
+
+	// configure session store for meet-director login
+	store := cookie.NewStore([]byte("secret"))
+	store.Options(sessions.Options{
+		Path:     "/",
+		MaxAge:   86400, // 1 day
+		HttpOnly: true,
+		Secure:   true,
+	})
+	router.Use(sessions.Sessions("mySession", store))
+
+	// basic security headers
+	router.Use(func(c *gin.Context) {
+		c.Writer.Header().Set("X-Frame-Options", "ALLOW-FROM https://referee-lights.michaelkingston.com.au")
+		c.Next()
+	})
+
+	// disable HTTP caching
+	router.Use(func(c *gin.Context) {
+		c.Writer.Header().Set("Cache-Control", "no-store, must-revalidate")
+		c.Writer.Header().Set("Pragma", "no-cache")
+		c.Writer.Header().Set("Expires", "0")
+		c.Next()
+	})
+
+	// health endpoint
+	router.GET("/health", controllers.Health)
+
+	// simple log endpoint
+	router.POST("/log", func(c *gin.Context) {
+		var payload struct {
+			Message string `json:"message"`
+			Level   string `json:"level"`
+		}
+		if err := c.ShouldBindJSON(&payload); err != nil {
+			logger.Warn.Printf("[SetupRouter /log] Invalid log payload: %v", err)
+			c.Status(http.StatusBadRequest)
+			return
+		}
+		switch payload.Level {
+		case "error":
+			logger.Error.Println(payload.Message)
+		case "warn":
+			logger.Warn.Println(payload.Message)
+		case "debug":
+			logger.Debug.Println(payload.Message)
+		default: // "info" + any unknown
+			logger.Info.Println(payload.Message)
+		}
+		c.Status(http.StatusOK)
+	})
+
+	// create services and controllers
+	occupancyService := services.NewOccupancyService()
+	positionController := controllers.NewPositionController(occupancyService)
+	meetDirectorController := controllers.NewAdminController(occupancyService, positionController)
+
+	// ------------------ public routes ------------------
+	router.GET("/", controllers.ShowMeets)
+	router.POST("/set-meet", controllers.SetMeetHandler)
+	router.GET("/login", controllers.PerformLogin)
+	router.POST("/login", controllers.LoginHandler)
+	router.GET("/left", controllers.Left)
+	router.GET("/center", controllers.Center)
+	router.GET("/right", controllers.Right)
+	router.GET("/referee/:meetName/:position", func(c *gin.Context) { controllers.RefereeHandler(c, occupancyService) })
+	router.GET("/heartbeat", func(c *gin.Context) { Handler(c.Writer, c.Request) })
+	router.SetHTMLTemplate(template.Must(template.ParseGlob("templates/*.html")))
+
+	// enforce “meetName in session” for any route that the meet director needs
+	router.Use(func(c *gin.Context) {
+		if strings.HasPrefix(c.Request.URL.Path, "/static/") {
+			c.Next()
+			return
+		}
+		if c.Request.URL.Path == "/login" || c.Request.URL.Path == "/referee-updates" {
+			c.Next()
+			return
+		}
+		if strings.HasPrefix(c.Request.URL.Path, "/referee/") ||
+			c.Request.URL.Path == "/left" ||
+			c.Request.URL.Path == "/center" ||
+			c.Request.URL.Path == "/right" {
+			c.Next()
+			return
+		}
+
+		session := sessions.Default(c)
+		if _, ok := session.Get("meetName").(string); !ok {
+			c.Redirect(http.StatusFound, "/")
+			c.Abort()
+			return
+		}
+		c.Next()
+	})
+
+	// ------------------ protected meet director routes ------------------
+	protected := router.Group("/")
+	protected.Use(middleware.AuthRequired)
+	protected.Use(func(c *gin.Context) {
+		session := sessions.Default(c)
+		if _, ok := session.Get("meetName").(string); !ok {
+			c.Redirect(http.StatusFound, "/")
+			c.Abort()
+			return
+		}
+		c.Next()
+	})
+	{
+		protected.GET("/index", controllers.Index)
+		protected.GET("/qrcode", controllers.GetQRCode)
+		protected.GET("/lights", controllers.Lights)
+		protected.GET("/occupancy", positionController.GetOccupancyAPI)
+		protected.POST("/position/vacate", positionController.VacatePosition)
+		protected.GET("/active-users", controllers.ActiveUsersHandler)
+		protected.GET("/admin-panel", meetDirectorController.AdminPanel)
+		protected.POST("/force-vacate", meetDirectorController.ForceVacate)
+		protected.POST("/reset-instance", meetDirectorController.ResetInstance)
+		protected.GET("/logout", func(c *gin.Context) { controllers.Logout(c, occupancyService) })
+		protected.POST("/logout", func(c *gin.Context) { controllers.Logout(c, occupancyService) })
+	}
+
+	// ------------------ sudo routes ------------------
+	sudoController := controllers.NewSudoController(occupancyService)
+	sudoRoutes := router.Group("/sudo")
+	{
+		sudoRoutes.Use(middleware.AuthRequired)
+		sudoRoutes.Use(middleware.SudoRequired())
+		{
+			sudoRoutes.GET("/", sudoController.SudoPanel)
+			sudoRoutes.POST("/force-vacate-ref", sudoController.ForceVacateRefForAnyMeet)
+			sudoRoutes.POST("/force-logout-meet-director", sudoController.ForceLogoutMeetDirector)
+			sudoRoutes.POST("/restart-meet", sudoController.RestartAndClearMeet)
+		}
+	}
+
+	// WebSocket route
+	router.GET("/referee-updates", func(c *gin.Context) {
+		websocket.ServeWs(c.Writer, c.Request)
+	})
+
+	// confirm template path
+	_, b, _, _ := runtime.Caller(0)
+	basePath := filepath.Dir(b)
+	templatesDir := filepath.Join(basePath, "../../templates")
+	if _, err := os.Stat(templatesDir); os.IsNotExist(err) {
+		log.Fatalf("[SetupRouter] Templates directory does not exist: %s", templatesDir)
+	}
+	router.SetHTMLTemplate(template.Must(template.ParseGlob(filepath.Join(templatesDir, "*.html"))))
+	logger.Debug.Printf("[SetupRouter] Templates Path: %s", templatesDir)
+	return router
+}
+
+var (
+	refereeSessions = make(map[string]time.Time)
+	sessionLock     = sync.Mutex{}
+)
+
+// Manager tracks active referees
+type Manager struct {
+	activeSessions map[string]time.Time
+	mu             sync.Mutex
+}
+
+// Handler updates the last seen timestamp of a referee
+func Handler(w http.ResponseWriter, r *http.Request) {
+	// Start a subsegment, but it may return nil if there's no parent segment
+	ctx, seg := xray.BeginSubsegment(r.Context(), "HeartbeatHandler")
+	if seg != nil {
+		defer seg.Close(nil)
+	}
+
+	// update request's context in case anything else reads it
+	r = r.WithContext(ctx)
+
+	refereeID := r.URL.Query().Get("referee_id")
+	if refereeID != "" && seg != nil {
+		_ = seg.AddAnnotation("refereeID", refereeID)
+	}
+
+	if refereeID == "" {
+		logger.Warn.Println("[Handler] Missing referee ID in query params")
+		http.Error(w, "Missing referee ID", http.StatusBadRequest)
+		return
+	}
+
+	sessionLock.Lock()
+	refereeSessions[refereeID] = time.Now()
+	sessionLock.Unlock()
+
+	logger.Debug.Printf("[Handler] Updated heartbeat for referee=%s at %v", refereeID, time.Now())
+
+	w.WriteHeader(http.StatusOK)
+	if _, err := fmt.Fprintln(w, "Heartbeat received"); err != nil {
+		logger.Warn.Printf("[Handler] Error writing response for referee=%s: %v", refereeID, err)
+	}
+}
+
+// NewHeartbeatManager initializes a heartbeat tracker
+func NewHeartbeatManager() *Manager {
+	return &Manager{
+		activeSessions: make(map[string]time.Time),
+	}
+}
+
+// UpdateHeartbeat marks a referee as active
+func (h *Manager) UpdateHeartbeat(refereeID string) {
+	h.mu.Lock()
+	defer h.mu.Unlock()
+	h.activeSessions[refereeID] = time.Now()
+	logger.Debug.Printf("[Manager.UpdateHeartbeat] Referee=%s updated at %v", refereeID, time.Now())
+}
+
+// CleanupInactiveSessions removes inactive referees
+func (h *Manager) CleanupInactiveSessions(timeout time.Duration) {
+	ticker := time.NewTicker(timeout)
+	go func() {
+		for range ticker.C {
+			h.mu.Lock()
+			for id, lastSeen := range h.activeSessions {
+				if time.Since(lastSeen) > timeout {
+					logger.Info.Printf("[Manager.CleanupInactiveSessions] Removing inactive referee=%s (timeout=%v)", id, timeout)
+					delete(h.activeSessions, id)
+				}
+			}
+			h.mu.Unlock()
+		}
+	}()
+}
+
+// CleanupRoutine removes referees that have been inactive
+//func CleanupRoutine() {
+//	ticker := time.NewTicker(10 * time.Second) // adjust interval as needed
+//	for range ticker.C {
+//		sessionLock.Lock()
+//		for id, lastSeen := range refereeSessions {
+//			if time.Since(lastSeen) > 1800*time.Second { // configurable timeout, 30 minutes
+//				logger.Info.Printf("[CleanupRoutine] Removing inactive referee=%s (30 minutes)", id)
+//				delete(refereeSessions, id)
+//			}
+//		}
+//		sessionLock.Unlock()
+//	}
+//}
 
 func main() {
 	// load env variables
@@ -68,7 +345,7 @@ func main() {
 	logger.Info.Println("[main] Starting application on port :8080")
 
 	// setup the router
-	router := app.SetupRouter(env)
+	router := SetupRouter(env)
 
 	// optional: Set X-Ray config
 	err = xray.Configure(xray.Config{
@@ -83,7 +360,7 @@ func main() {
 	xrayHandler := xray.Handler(xraySegmentNamer, router)
 
 	// start background routines
-	hbManager := heartbeat.NewHeartbeatManager()
+	hbManager := NewHeartbeatManager()
 	go hbManager.CleanupInactiveSessions(30 * time.Second)
 	go websocket.HandleMessages()
