+++ controllers/page_controller.go
@@ -3,14 +3,17 @@
 package controllers
 
 import (
+	"encoding/json"
 	"fmt"
-	"github.com/aws/aws-xray-sdk-go/xray"
 	"net/http"
+	"os"
 	"sync"
 
+	"github.com/aws/aws-xray-sdk-go/xray"
 	"github.com/gin-contrib/sessions"
 	"github.com/gin-gonic/gin"
 	"github.com/skip2/go-qrcode"
+
 	"go-ref-lights/logger"
 	"go-ref-lights/models"
 	"go-ref-lights/services"
@@ -385,4 +388,53 @@ func renderLeft(c *gin.Context, meetName string) {
 		"meetName":     meetName,
 	}
 	c.HTML(http.StatusOK, "left.html", data)
+}
+
+// --------------- global variables ---------------
+
+// loadMeetsFunc allows dependency injection for testing.
+var loadMeetsFunc = LoadMeets
+
+// ------------- meet configuration management -------------
+
+// LoadMeets loads the meet configuration from `./config/meets.json`.
+// this function retrieves the available meets and their details from the JSON file.
+func LoadMeets() (*models.MeetCreds, error) {
+	meetsPath := "config/meets.json" // #nosec G101
+	if env := os.Getenv("MEETS_PATH"); env != "" {
+		meetsPath = env
+	}
+	// read the config file
+	data, err := os.ReadFile(meetsPath)
+	if err != nil {
+		return nil, err
+	}
+
+	var meets models.MeetCreds
+	if err := json.Unmarshal(data, &meets); err != nil {
+		return nil, err
+	}
+
+	logger.Info.Printf("[LoadMeets] Successfully loaded %d meets", len(meets.Meets))
+	return &meets, nil
+}
+
+// -------------- meet selection handling --------------
+
+// ShowMeets renders the meet selection page.
+// it fetches the list of available meets and passes them to the template.
+// if loading fails, it returns an HTTP 500 response.
+func ShowMeets(c *gin.Context) {
+	// retrieve meet data using a mockable function for easier testing
+	meetsData, err := loadMeetsFunc()
+	if err != nil {
+		logger.Error.Printf("[ShowMeets] Failed to load meets: %v", err)
+		c.String(http.StatusInternalServerError, "Failed to load meets")
+		return
+	}
+
+	// render the meet selection page with available meets
+	c.HTML(http.StatusOK, "choose_meet.html", gin.H{
+		"availableMeets": meetsData.Meets,
+	})
 }
